added keywords to models

+++ creator.py
@@ -1,14 +1,15 @@
 # coding=utf-8
 import json
+import os
 
 from models import Owner, Account, Client
 
 owner = Owner('Sample Owner Company', '43-190 Mikołów\nul. Fajna 66', 1111111111,
-              Account('Cool bank', '33 5555 5555 5555 5555 5555 5555'))
+              Account('Cool bank', '33 5555 5555 5555 5555 5555 5555'), 'sample')
 
-client = Client('Sample client company', '43-190 Mikołów\nul. Fajna 77', 2222222222)
+client = Client('Sample client company', '43-190 Mikołów\nul. Fajna 77', 60, 2222222222, 'sample')
 
-deliver_to = Client('Sample deliver company', '43-190 Mikołów\nul. Fajna 77', 3333333333)
+deliver_to = Client('Sample deliver company', '43-190 Mikołów\nul. Fajna 77', 0,3333333333, 'sampleDeliver')
 
 
 def to_json(complex_object):
@@ -17,8 +18,9 @@ def to_json(complex_object):
 
 
 def create(file_name, source):
-    with open(file_name, 'w') as new_file:
-        new_file.write(to_json(source.__dict__))
+    if not os.path.isfile(file_name):
+        with open(file_name, 'w') as new_file:
+            new_file.write(to_json(source.__dict__))
 
 
 create('owner.json', owner)
